+++ example/lib/sources.dart
@@ -1,0 +1,11 @@
+final videoSources = [
+  'asset://assets/01.mp4',
+  'asset://assets/02.mp4',
+  'asset://assets/03.mp4',
+  'asset://assets/04.mp4',
+  'asset://assets/05.mp4',
+  'asset://assets/06.mp4',
+  'https://flutter.github.io/assets-for-api-docs/assets/videos/bee.mp4',
+  'https://devstreaming-cdn.apple.com/videos/streaming/examples/bipbop_4x3/bipbop_4x3_variant.m3u8',
+  'https://download.samplelib.com/mp3/sample-3s.mp3',
+];